Rename Text->TextTool
