+++ src/lib/components/Tools/TextTool.svelte
